refactor: Server, ServerService, JCFServerService 구조 변경
