+++ src/main/java/com/sprint/mission/discodeit/composit/Channel.java
@@ -3,7 +3,6 @@ package com.sprint.mission.discodeit.composit;
 import com.sprint.mission.discodeit.Iterator.Iterator;
 import com.sprint.mission.discodeit.entity.Message;
 
-import java.util.Date;
 import java.util.LinkedList;
 import java.util.Scanner;
