+++ src/main/java/com/sprint/mission/discodeit/service/ServerService.java
@@ -1,9 +1,8 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.composit.Channel;
-import com.sprint.mission.discodeit.entity.User;
-
 public interface ServerService {
-
-
+    public abstract void print();
+    public abstract void addChannel(String name);
+    public abstract void update(String targetName, String replaceName);
+    public abstract void remove(String targetName);
 }
