+++ src/main/java/com/sprint/mission/discodeit/service/UserService.java
@@ -1,7 +1,5 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.entity.User;
-
 public interface UserService {
     public abstract void currentHead();
     public abstract void createServer(String name);
