+++ src/main/java/com/sprint/mission/discodeit/service/jcf/JCFMessageService.java
@@ -1,10 +1,5 @@
 package com.sprint.mission.discodeit.service.jcf;
 
-import com.sprint.mission.discodeit.entity.Message;
-import com.sprint.mission.discodeit.service.MessageService;
-
-import java.util.Scanner;
-
 public class JCFMessageService {
 
 }
